Added usage in Documentation and some few corrections to the code.
