+++ wattpad_scraper/utils/__init__.py
@@ -1,2 +1,3 @@
-from .request import get, ReadingList
+from .request import get
+from .reading_list import ReadingList
 from .log import Log
